Creation of Graph components (Pie and bar)
